# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MCP Tool Evaluation System API
  description: |
    REST API for evaluating Model Context Protocol (MCP) tool effectiveness.
    
    **Features**:
    - Tool selection evaluation with LLM providers
    - Parameter validation (completeness, correctness, type conformance)
    - Inter-tool similarity detection and recommendations
    - Comprehensive metrics (precision, recall, F1, parameter accuracy)
    - Security-first architecture (runtime API keys, never persisted)
    - Fresh tool ingestion per test run for accuracy
    
    **Generated**: 2025-11-10
    **Git Commit**: 859d728254fc64586c05039908056c6b4bec1709
  version: 2.0.0
  contact:
    name: mcp-tef Team
    url: https://github.com/StacklokLabs/mcp-tef
servers:
  - url: https://localhost:8000
    description: Local development server (HTTPS with auto-generated cert)
  - url: http://localhost:8000
    description: Local development server (HTTP, use --no-tls-enabled)
...
paths:
...
  /test-cases:
    post:
      tags: [test-cases]
      summary: Create test case
      description: |
        **003-remove-api-key**: No model_id required. Model config provided at execution time.
      operationId: createTestCase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestCaseCreate'
      responses:
        '201':
          description: Test case created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestCaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'

    get:
      tags: [test-cases]
      summary: List all test cases
      operationId: listTestCases
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: List of test cases
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TestCaseResponse'
                  total:
                    type: integer
  /test-cases/{test_case_id}:
    get:
      tags: [test-cases]
      summary: Get test case details
      operationId: getTestCase
      parameters:
        - $ref: '#/components/parameters/TestCaseIdParam'
      responses:
        '200':
          description: Test case details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestCaseResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [test-cases]
      summary: Delete test case
      operationId: deleteTestCase
      parameters:
        - $ref: '#/components/parameters/TestCaseIdParam'
      responses:
        '204':
          description: Test case deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # TEST RUNS (All features: 001 base, 003 runtime API key, 004 tool ingestion)
  # ============================================================================
  /test-cases/{test_case_id}/run:
    post:
      tags: [test-runs]
      summary: Execute test case
      description: |
        Execute a test case with runtime model configuration and API key.
        
        **003-remove-api-key**: API key via header, model settings in body.
        **004-decouple-tool-ingestion**: Tools ingested fresh from all associated servers (1-5s per server overhead).
      operationId: runTestCase
      parameters:
        - $ref: '#/components/parameters/TestCaseIdParam'
        - name: X-Model-API-Key
          in: header
          required: false
          description: |
            LLM provider API key (not persisted). Required for cloud providers (openai, anthropic, openrouter).
            Optional for local providers (ollama).
          schema:
            type: string
          example: "sk-..."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestRunExecuteRequest'
      responses:
        '200':
          description: Test run completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestRunResponse'
        '401':
          description: Missing or invalid API key
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          description: Tool ingestion failed or LLM error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestRunResponse'
  /test-runs/{test_run_id}:
    get:
      tags: [test-runs]
      summary: Get test run details
      operationId: getTestRun
      parameters:
        - $ref: '#/components/parameters/TestRunIdParam'
      responses:
        '200':
          description: Test run details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestRunResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
...
    TestCaseIdParam:
      name: test_case_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TestRunIdParam:
      name: test_run_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 100
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
  schemas:
...
    TestCaseCreate:
      type: object
      required: [name, query, expected_mcp_server_name, expected_tool_name]
      properties:
        name:
          type: string
          minLength: 1
        query:
          type: string
          minLength: 1
        expected_mcp_server_name:
          type: string
          minLength: 1
        expected_tool_name:
          type: string
          minLength: 1
        expected_parameters:
          type: object
          nullable: true
    TestCaseResponse:
      type: object
      required: [id, name, query, expected_mcp_server_name, expected_tool_name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        query:
          type: string
        expected_mcp_server_name:
          type: string
        expected_tool_name:
          type: string
        expected_parameters:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ==========================================================================
    # MODEL SETTINGS (003-remove-api-key)
    # ==========================================================================
    ModelSettingsCreate:
      type: object
      required: [provider, model]
      properties:
        provider:
          type: string
          minLength: 1
          maxLength: 100
          example: "openai"
        model:
          type: string
          minLength: 1
          maxLength: 100
          example: "gpt-4"
        timeout:
          type: integer
          default: 30
          minimum: 1
          maximum: 300
        temperature:
          type: number
          format: float
          default: 0.4
          minimum: 0.0
          maximum: 2.0
        max_retries:
          type: integer
          default: 3
          minimum: 0
          maximum: 10
    ModelSettingsResponse:
      type: object
      required: [id, provider, model, timeout, temperature, max_retries, created_at]
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
        model:
          type: string
        timeout:
          type: integer
        temperature:
          type: number
          format: float
        max_retries:
          type: integer
        created_at:
          type: string
          format: date-time

    # ==========================================================================
    # TEST RUNS (All features integrated)
    # ==========================================================================
    TestRunExecuteRequest:
      type: object
      required: [model_config]
      properties:
        model_config:
          $ref: '#/components/schemas/ModelSettingsCreate'
    TestRunResponse:
      type: object
      required: [id, test_case_id, model_settings, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        test_case_id:
          type: string
          format: uuid
        model_settings:
          $ref: '#/components/schemas/ModelSettingsResponse'
          description: '003: Embedded model settings (API key NOT included)'
        status:
          type: string
          enum: [pending, running, completed, failed]
        confidence_score:
          type: number
          format: float
          nullable: true
          minimum: 0.0
          maximum: 1.0
        classification:
          type: string
          enum: [TP, FP, TN, FN]
          nullable: true
        execution_time_ms:
          type: integer
          nullable: true
        error_message:
          type: string
          nullable: true
          description: '004: Includes tool ingestion errors'
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true

    # ==========================================================================
    # EVALUATION RESULTS (001-this-is-an)
    # ==========================================================================
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security: []